move allocation fxns to seperate py script
