+++ flowsa/flowbyfunctions.py
@@ -2,7 +2,6 @@
 Helper functions for flowbyactivity and flowbysector data
 """
 
-import flowsa
 import pandas as pd
 import numpy as np
 from flowsa.common import log, get_county_FIPS, get_state_FIPS, US_FIPS, activity_fields, \
@@ -289,197 +288,6 @@ def harmonize_units(df):
     return df
 
 
-def allocate_by_sector(df_w_sectors, source_name, allocation_source, allocation_method, group_cols, **kwargs):
-    """
-    Create an allocation ratio for df
-
-    :param df_w_sectors: df with column of sectors
-    :param source_name: the name of the FBA df being allocated
-    :param allocation_source: The name of the FBA allocation dataframe
-    :param allocation_method: currently written for 'proportional'
-    :param group_cols: columns on which to base aggregation and disaggregation
-    :return: df with FlowAmountRatio for each sector
-    """
-
-    # first determine if there is a special case with how the allocation ratios are created
-    if allocation_method == 'proportional-flagged':
-        # if the allocation method is flagged, subset sectors that are flagged/notflagged, where nonflagged sectors \
-        # have flowamountratio=1
-        if kwargs != {}:
-            if 'flowSubsetMapped' in kwargs:
-                fsm = kwargs['flowSubsetMapped']
-                flagged = fsm[fsm['disaggregate_flag'] == 1]
-                flagged_names = flagged['SectorConsumedBy'].tolist()
-
-                nonflagged = fsm[fsm['disaggregate_flag'] == 0]
-                nonflagged_names = nonflagged['SectorConsumedBy'].tolist()
-
-                # subset the original df so rows of data that run through the proportioanl allocation process are
-                # sectors included in the flagged list
-                df_w_sectors_nonflagged = df_w_sectors.loc[
-                    (df_w_sectors[fbs_activity_fields[0]].isin(nonflagged_names)) |
-                    (df_w_sectors[fbs_activity_fields[1]].isin(nonflagged_names))
-                    ].reset_index(drop=True)
-                df_w_sectors_nonflagged = df_w_sectors_nonflagged.assign(FlowAmountRatio=1)
-
-                df_w_sectors = df_w_sectors.loc[(df_w_sectors[fbs_activity_fields[0]].isin(flagged_names)) |
-                                                (df_w_sectors[fbs_activity_fields[1]].isin(flagged_names))
-                                                ].reset_index(drop=True)
-            else:
-                log.error('The proportional-flagged allocation method requires a column "disaggregate_flag" in the'
-                          'flow_subset_mapped df')
-
-    # run sector aggregation fxn to determine total flowamount for each level of sector
-    if len(df_w_sectors) == 0:
-        allocation_df = df_w_sectors_nonflagged.copy()
-    else:
-        df1 = sector_aggregation(df_w_sectors, group_cols)
-        # run sector disaggregation to capture one-to-one naics4/5/6 relationships
-        df2 = sector_disaggregation(df1, group_cols)
-
-        # if statements for method of allocation
-        # either 'proportional' or 'proportional-flagged'
-        allocation_df = []
-        if allocation_method == 'proportional' or allocation_method == 'proportional-flagged':
-            allocation_df = proportional_allocation_by_location(df2)
-        else:
-            log.error('Must create function for specified method of allocation')
-
-        if allocation_method == 'proportional-flagged':
-            # drop rows where values are not in flagged names
-            allocation_df = allocation_df.loc[(allocation_df[fbs_activity_fields[0]].isin(flagged_names)) |
-                                              (allocation_df[fbs_activity_fields[1]].isin(flagged_names))
-                                              ].reset_index(drop=True)
-            # concat the flagged and nonflagged dfs
-            allocation_df = pd.concat([allocation_df, df_w_sectors_nonflagged],
-                                      ignore_index=True).sort_values(['SectorProducedBy', 'SectorConsumedBy'])
-
-    return allocation_df
-
-
-def proportional_allocation_by_location(df):
-    """
-    Creates a proportional allocation based on all the most aggregated sectors within a location
-    Ensure that sectors are at 2 digit level - can run sector_aggregation() prior to using this function
-    :param df:
-    :param sectorcolumn:
-    :return:
-    """
-
-    # tmp drop NoneType
-    df = replace_NoneType_with_empty_cells(df)
-
-    # find the shortest length sector
-
-    denom_df = df.loc[(df['SectorProducedBy'].apply(lambda x: len(x) == 2)) |
-                      (df['SectorConsumedBy'].apply(lambda x: len(x) == 2))]
-    denom_df = denom_df.assign(Denominator=denom_df['FlowAmount'].groupby(
-        denom_df['Location']).transform('sum'))
-    denom_df_2 = denom_df[['Location', 'LocationSystem', 'Year', 'Denominator']].drop_duplicates()
-    # merge the denominator column with fba_w_sector df
-    allocation_df = df.merge(denom_df_2, how='left')
-    # calculate ratio
-    allocation_df.loc[:, 'FlowAmountRatio'] = allocation_df['FlowAmount'] / allocation_df[
-        'Denominator']
-    allocation_df = allocation_df.drop(columns=['Denominator']).reset_index()
-
-    # add nonetypes
-    allocation_df = replace_strings_with_NoneType(allocation_df)
-
-    return allocation_df
-
-
-def proportional_allocation_by_location_and_activity(df, sectorcolumn):
-    """
-    Creates a proportional allocation within each aggregated sector within a location
-    :param df:
-    :param sectorcolumn:
-    :return:
-    """
-
-    # tmp replace NoneTypes with empty cells
-    df = replace_NoneType_with_empty_cells(df)
-
-    # denominator summed from highest level of sector grouped by location
-    short_length = min(df[sectorcolumn].apply(lambda x: len(str(x))).unique())
-    # want to create denominator based on short_length
-    denom_df = df.loc[df[sectorcolumn].apply(lambda x: len(x) == short_length)].reset_index(drop=True)
-    grouping_cols = [e for e in ['Location', 'Activity', 'ActivityConsumedBy', 'ActivityProducedBy']
-                     if e in denom_df.columns.values.tolist()]
-    denom_df.loc[:, 'Denominator'] = denom_df.groupby(grouping_cols)['HelperFlow'].transform('sum')
-
-    # list of column headers, that if exist in df, should be aggregated using the weighted avg fxn
-    possible_column_headers = ('Location', 'LocationSystem', 'Year', 'Activity', 'ActivityConsumedBy', 'ActivityProducedBy')
-    # list of column headers that do exist in the df being aggregated
-    column_headers = [e for e in possible_column_headers if e in denom_df.columns.values.tolist()]
-    merge_headers = column_headers.copy()
-    column_headers.append('Denominator')
-    # create subset of denominator values based on Locations and Activities
-    denom_df_2 = denom_df[column_headers].drop_duplicates().reset_index(drop=True)
-    # merge the denominator column with fba_w_sector df
-    allocation_df = df.merge(denom_df_2,
-                             how='left',
-                             left_on=merge_headers,
-                             right_on=merge_headers)
-    # calculate ratio
-    allocation_df.loc[:, 'FlowAmountRatio'] = allocation_df['HelperFlow'] / allocation_df['Denominator']
-    allocation_df = allocation_df.drop(columns=['Denominator']).reset_index(drop=True)
-
-    # fill empty cols with NoneType
-    allocation_df = replace_strings_with_NoneType(allocation_df)
-    # fill na values with 0
-    allocation_df['HelperFlow'] = allocation_df['HelperFlow'].fillna(0)
-
-    return allocation_df
-
-
-
-# def proportional_allocation_by_location_and_sector(df, sectorcolumn):
-#     """
-#     Creates a proportional allocation within each aggregated sector within a location
-#     :param df:
-#     :param sectorcolumn:
-#     :return:
-#     """
-#     from flowsa.common import load_source_catalog
-#
-#     cat = load_source_catalog()
-#     src_info = cat[pd.unique(df['SourceName'])[0]]
-#     # load source catalog to determine the level of sector aggregation associated with a crosswalk
-#     level_of_aggregation = src_info['sector_aggregation_level']
-#
-#     # denominator summed from highest level of sector grouped by location
-#     short_length = min(df[sectorcolumn].apply(lambda x: len(str(x))).unique())
-#     # want to create denominator based on short_length - 1, unless short_length = 2
-#     denom_df = df.loc[df[sectorcolumn].apply(lambda x: len(x) == short_length)]
-#     if (level_of_aggregation == 'disaggregated') & (short_length != 2):
-#         short_length = short_length - 1
-#         denom_df.loc[:, 'sec_tmp'] = denom_df[sectorcolumn].apply(lambda x: x[0:short_length])
-#         denom_df.loc[:, 'Denominator'] = denom_df.groupby(['Location', 'sec_tmp'])['FlowAmount'].transform('sum')
-#     else:  # short_length == 2:]
-#         denom_df.loc[:, 'Denominator'] = denom_df['FlowAmount']
-#         denom_df.loc[:, 'sec_tmp'] = denom_df[sectorcolumn]
-#     # if short_length == 2:
-#     #     denom_df.loc[:, 'Denominator'] = denom_df['FlowAmount']
-#     #     denom_df.loc[:, 'sec_tmp'] = denom_df[sectorcolumn]
-#     # else:
-#     #     short_length = short_length - 1
-#     #     denom_df.loc[:, 'sec_tmp'] = denom_df[sectorcolumn].apply(lambda x: x[0:short_length])
-#     #     denom_df.loc[:, 'Denominator'] = denom_df.groupby(['Location', 'sec_tmp'])['FlowAmount'].transform('sum')
-#
-#     denom_df_2 = denom_df[['Location', 'LocationSystem', 'Year', 'sec_tmp', 'Denominator']].drop_duplicates()
-#     # merge the denominator column with fba_w_sector df
-#     df.loc[:, 'sec_tmp'] = df[sectorcolumn].apply(lambda x: x[0:short_length])
-#     allocation_df = df.merge(denom_df_2, how='left', left_on=['Location', 'LocationSystem', 'Year', 'sec_tmp'],
-#                              right_on=['Location', 'LocationSystem', 'Year', 'sec_tmp'])
-#     # calculate ratio
-#     allocation_df.loc[:, 'FlowAmountRatio'] = allocation_df['FlowAmount'] / allocation_df[
-#         'Denominator']
-#     allocation_df = allocation_df.drop(columns=['Denominator', 'sec_tmp']).reset_index(drop=True)
-#
-#     return allocation_df
-
-
 def sector_ratios(df, sectorcolumn):
     """
     Determine ratios of the less aggregated sectors within a more aggregated sector
@@ -517,153 +325,6 @@ def sector_ratios(df, sectorcolumn):
     return df_w_ratios
 
 
-def allocation_helper(df_w_sector, attr, method, v):
-    """
-    Used when two df required to create allocation ratio
-    :param df_w_sector:
-    :param method: currently written for 'multiplication' and 'proportional'
-    :param attr:
-    :return:
-    """
-
-    from flowsa.data_source_scripts.Blackhurst_IO import scale_blackhurst_results_to_usgs_values
-    from flowsa.data_source_scripts.BLS_QCEW import clean_bls_qcew_fba, bls_clean_allocation_fba_w_sec
-    from flowsa.mapping import add_sectors_to_flowbyactivity
-
-    helper_allocation = flowsa.getFlowByActivity(datasource=attr['helper_source'],
-                                                 year=attr['helper_source_year'],
-                                                 flowclass=attr['helper_source_class'])
-    # clean df
-    helper_allocation = clean_df(helper_allocation, flow_by_activity_fields, fba_fill_na_dict)
-    helper_allocation = harmonize_units(helper_allocation)
-
-    # agg data if necessary or filter
-    # determine to scale
-    to_scale = min(fips_number_key[attr['allocation_from_scale']], fips_number_key[v['geoscale_to_use']])
-    if fips_number_key[attr['helper_from_scale']] > to_scale:
-        helper_allocation = agg_by_geoscale(helper_allocation,
-                                            attr['helper_from_scale'],
-                                            list(fips_number_key.keys())[list(fips_number_key.values()).index(to_scale)],
-                                            fba_default_grouping_fields)
-    else:
-        helper_allocation = filter_by_geoscale(helper_allocation, attr['helper_from_scale'])
-
-    if 'clean_helper_fba' in attr:
-        log.info("Cleaning " + attr['helper_source'] + ' FBA')
-        # tmp hard coded - need to generalize
-        if attr['helper_source'] == 'BLS_QCEW':
-            helper_allocation = clean_bls_qcew_fba(helper_allocation, attr=attr)
-            # helper_allocation = getattr(sys.modules[__name__], attr["clean_helper_fba"])(helper_allocation, attr)
-
-    # drop rows with flowamount = 0
-    helper_allocation = helper_allocation[helper_allocation['FlowAmount'] != 0]
-
-    # assign naics to allocation dataset
-    helper_allocation = add_sectors_to_flowbyactivity(helper_allocation,
-                                                      sectorsourcename=method['target_sector_source'])
-    # clean up helper fba with sec
-    if 'clean_helper_fba_wsec' in attr:
-        log.info("Cleaning " + attr['helper_source'] + ' FBA with sectors')
-        # tmp hard coded - need to generalize
-        if attr['helper_source'] == 'BLS_QCEW':
-            helper_allocation = bls_clean_allocation_fba_w_sec(helper_allocation, attr=attr, method=method)
-            # helper_allocation = getattr(sys.modules[__name__], attr["clean_helper_fba_wsec"])(helper_allocation, attr, method)
-
-    # run sector disagg to capture any missing lower level naics
-    helper_allocation = sector_disaggregation(helper_allocation, fba_mapped_default_grouping_fields)
-
-    # generalize activity field names to enable link to water withdrawal table
-    helper_allocation = collapse_activity_fields(helper_allocation)
-    # drop any rows not mapped
-    helper_allocation = helper_allocation[helper_allocation['Sector'].notnull()]
-    # drop columns
-    helper_allocation = helper_allocation.drop(columns=['Activity', 'Min', 'Max'])
-
-    # rename column
-    helper_allocation = helper_allocation.rename(columns={"FlowAmount": 'HelperFlow'})
-
-    # determine the df_w_sector column to merge on
-    df_w_sector = replace_strings_with_NoneType(df_w_sector)
-    sec_consumed_list = df_w_sector['SectorConsumedBy'].drop_duplicates().values.tolist()
-    sec_produced_list = df_w_sector['SectorProducedBy'].drop_duplicates().values.tolist()
-    # if a sector field column is not all 'none', that is the column to merge
-    if all(v is None for v in sec_consumed_list):
-        sector_col_to_merge = 'SectorProducedBy'
-    elif all(v is None for v in sec_produced_list):
-        sector_col_to_merge = 'SectorConsumedBy'
-    else:
-        log.error('There is not a clear sector column to base merge with helper allocation dataset')
-
-    # merge allocation df with helper df based on sectors, depending on geo scales of dfs
-    if (attr['helper_from_scale'] == 'state') and (attr['allocation_from_scale'] == 'county'):
-        helper_allocation.loc[:, 'Location_tmp'] = helper_allocation['Location'].apply(lambda x: x[0:2])
-        df_w_sector.loc[:, 'Location_tmp'] = df_w_sector['Location'].apply(lambda x: x[0:2])
-        # merge_columns.append('Location_tmp')
-        modified_fba_allocation = df_w_sector.merge(helper_allocation[['Location_tmp', 'Sector', 'HelperFlow']],
-                                                    how='left',
-                                                    left_on=['Location_tmp', sector_col_to_merge],
-                                                    right_on=['Location_tmp', 'Sector'])
-        modified_fba_allocation = modified_fba_allocation.drop(columns=['Location_tmp'])
-    elif (attr['helper_from_scale'] == 'national') and (attr['allocation_from_scale'] != 'national'):
-        modified_fba_allocation = df_w_sector.merge(helper_allocation[['Sector', 'HelperFlow']],
-                                                    how='left',
-                                                    left_on=[sector_col_to_merge],
-                                                    right_on=['Sector'])
-    else:
-        modified_fba_allocation = df_w_sector.merge(helper_allocation[['Location', 'Sector', 'HelperFlow']],
-                                                    how='left',
-                                                    left_on=['Location', sector_col_to_merge],
-                                                    right_on=['Location', 'Sector'])
-
-    # modify flow amounts using helper data
-    if 'multiplication' in attr['helper_method']:
-        # todo: modify so if missing data, replaced with value from one geoscale up instead of national
-        # todo: modify year after merge if necessary
-        # if missing values (na or 0), replace with national level values
-        replacement_values = helper_allocation[helper_allocation['Location'] == US_FIPS].reset_index(
-            drop=True)
-        replacement_values = replacement_values.rename(columns={"HelperFlow": 'ReplacementValue'})
-        modified_fba_allocation = modified_fba_allocation.merge(
-            replacement_values[['Sector', 'ReplacementValue']], how='left')
-        modified_fba_allocation.loc[:, 'HelperFlow'] = modified_fba_allocation['HelperFlow'].fillna(
-            modified_fba_allocation['ReplacementValue'])
-        modified_fba_allocation.loc[:, 'HelperFlow'] = np.where(modified_fba_allocation['HelperFlow'] == 0,
-                                                                modified_fba_allocation['ReplacementValue'],
-                                                                modified_fba_allocation['HelperFlow'])
-
-        # replace non-existent helper flow values with a 0, so after multiplying, don't have incorrect value associated
-        # with new unit
-        modified_fba_allocation['HelperFlow'] = modified_fba_allocation['HelperFlow'].fillna(value=0)
-        modified_fba_allocation.loc[:, 'FlowAmount'] = modified_fba_allocation['FlowAmount'] * \
-                                                       modified_fba_allocation['HelperFlow']
-        # drop columns
-        modified_fba_allocation = modified_fba_allocation.drop(columns=["HelperFlow", 'ReplacementValue', 'Sector'])
-
-    elif attr['helper_method'] == 'proportional':
-        modified_fba_allocation = proportional_allocation_by_location_and_activity(modified_fba_allocation,
-                                                                                   sector_col_to_merge)
-        modified_fba_allocation['FlowAmountRatio'] = modified_fba_allocation['FlowAmountRatio'].fillna(0)
-        modified_fba_allocation.loc[:, 'FlowAmount'] = modified_fba_allocation['FlowAmount'] * \
-                                                       modified_fba_allocation['FlowAmountRatio']
-        modified_fba_allocation = modified_fba_allocation.drop(columns=['FlowAmountRatio', 'HelperFlow', 'Sector'])
-
-    # drop rows of 0
-    modified_fba_allocation = modified_fba_allocation[modified_fba_allocation['FlowAmount'] != 0].reset_index(drop=True)
-
-    # todo: change units
-    modified_fba_allocation.loc[modified_fba_allocation['Unit'] == 'gal/employee', 'Unit'] = 'gal'
-
-    # option to scale up fba values
-    if 'scaled' in attr['helper_method']:
-        log.info("Scaling " + attr['helper_source'] + ' to FBA values')
-        # tmp hard coded - need to generalize
-        if attr['helper_source'] == 'BLS_QCEW':
-            modified_fba_allocation = scale_blackhurst_results_to_usgs_values(modified_fba_allocation, attr)
-            # modified_fba_allocation = getattr(sys.modules[__name__], attr["scale_helper_results"])(modified_fba_allocation, attr)
-
-    return modified_fba_allocation
-
-
 def sector_aggregation(df, group_cols):
     """
     Function that checks if a sector length exists, and if not, sums the less aggregated sector
@@ -672,6 +333,7 @@ def sector_aggregation(df, group_cols):
     :return:
     """
 
+    # todo: add statement that if sector-like, drop activity cols if exist (then add them back as copies of sector cols)
     # ensure None values are not strings
     df = replace_NoneType_with_empty_cells(df)
